Fix the response URL, remove $downpomf
There are certain Pomf clones which violate the API and return only the
short filename in the "url" response. The intended functionality of the
url is the full URL, to be documented.

Pantsu.cat returns the full URL as intended, and therefore this script
doesn't require to know from what subdomains the files come from. The
server API is responsible for informing the client where to find the
file from. It just works.

This also fixes a ShellCheck error "SC2016: Expressions don't expand in
single quotes, use double quotes for that.".